refactor: extract worktree cleanup engine
